requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
libsql-experimental==0.0.55
sqlalchemy-libsql==0.2.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.4.0
build==1.3.0
certifi==2025.8.3
click==8.2.1
dateparser==1.2.2
decorator==5.2.1
executing==2.2.1
fastapi==0.116.1
frozenlist==1.8.0
google-genai==1.0.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
multidict==6.7.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pip-tools==7.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2025.8.29
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.47.3
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.35.0
wcwidth==0.2.13
wheel==0.45.1
yarl==1.22.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
beautifulsoup4
apscheduler>=3.10.0
